rag/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
google-generativeai
langchain
langchain-community
langchain-google-genai
langchain-qdrant
tiktoken
unstructured
markdown
psycopg2-binary
openai
openai-agents
sentence-transformers